+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * Bugfix: Adhearsion call actors should do recursion detection on #inspect
   * Bugfix: Fix controller block context on JRuby
   * Bugfix: Require only Bundler groups appropriate to application environment
+  * Bugfix: Call event handlers now cannot crash the call, and no longer stop on return value
 
 # [2.4.0](https://github.com/adhearsion/adhearsion/compare/v2.3.5...v2.4.0) - [2013-08-29](https://rubygems.org/gems/adhearsion/versions/2.4.0)
   * Deprecation: Ruby 1.9.2 support is deprecated and will be dropped in a future version of Adhearsion